requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# ETL orchestration
prefect>=2.10.0

# Data processing and analysis
dask[dataframe]>=2023.1.0
dask[distributed]>=2023.1.0

# Date and time utilities
python-dateutil>=2.8.0

# File I/O
openpyxl>=3.1.0
xlrd>=2.0.0

# Data validation
pydantic>=2.0.0

# Logging and monitoring
structlog>=23.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Optional: For advanced analytics
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Optional: For cloud deployment
boto3>=1.26.0
azure-storage-blob>=12.0.0
